Use size_t in strrcmp().
git-svn-id: https://svn.xiph.org/trunk/ezstream@12616 0101bb08-14d6-0310-b084-bc0e0c8e3800

+++ src/ezstream.c
@@ -167,8 +167,8 @@ basename(const char *fileName)
 int
 strrcmp(const char *s, const char *sub)
 {
-	int	slen = strlen(s);
-	int	sublen = strlen(sub);
+	size_t	slen = strlen(s);
+	size_t	sublen = strlen(sub);
 
 	if (sublen > slen)
 		return (1);
